Rename package: routes -> pages
