+++ optaweb-vehicle-routing-frontend/src/ui/pages/Vehicles.tsx
